Backport to 1.20.6

+++ StaffMod/src/main/kotlin/opekope2/avm_staff/internal/staff/handler/CampfireHandler.kt
@@ -158,7 +158,7 @@ internal class CampfireHandler(
     data class Properties(
         val nonFlammableBlockFireChance: Double,
         val flammableBlockFireChance: Double,
-        val attackFireSeconds: Float,
+        val attackFireSeconds: Int,
         val flameFireTicks: Int,
         val rocketThrust: Double
     )
